% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allocation_Distance.R
\name{Allocation_Distance}
\alias{Allocation_Distance}
\title{Allocation Distance}
\usage{
Allocation_Distance(Z, Xr)
}
\arguments{
\item{Z}{dataset}

\item{Xr}{Center coordinates}
}
\value{
List with distances to center and allocation matrix
}
\description{
\code{Khorshidi et al., 2019}.
}
\examples{
doMIsaul:::Allocation_Distance(iris[, 1:4], iris[c(1,5, 10), 1:4])
}
