% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovatab.R
\name{anovatab}
\alias{anovatab}
\title{Table of ANOVA test for several explanatory variables}
\usage{
anovatab(data, main.var, var, nb.dec = 1)
}
\arguments{
\item{data}{The database.}

\item{main.var}{string. The categorical variable against the variable
will be compared.}

\item{var}{strings. The categorical variable to compare with.}

\item{nb.dec}{digit. Number of decimals for the mean and quantiles}
}
\value{
table with mean and q1 q3 for each level and p-value of ANOVA test.
}
\description{
Performs the function \code{aov()} for the \code{main.var} and \code{var}.
}
\keyword{internal}
