% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH_index_functions.R
\name{CH}
\alias{CH}
\title{\code{CH} index}
\usage{
CH(data, k, method)
}
\arguments{
\item{data}{dataframe for which the number of cluster should be estimated.}

\item{k}{integer, number of clusters.}

\item{method}{string, clustering algorithm to use. Only available value for
know is "\code{kmed}" for \code{K-medians} clustering.}
}
\value{
a list, containing the criterion value and the partition.
}
\description{
Computes CH index for the selected number of clusters and clustering
algorithm.
}
\examples{
doMIsaul:::CH(iris[, 1:4], 5, "kmed")
}
\keyword{internal}
