% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2tab.R
\name{chi2tab}
\alias{chi2tab}
\title{Table of chisq.test test for several explanotory varaibles}
\usage{
chi2tab(data, main.var, var, nb.dec = 1)
}
\arguments{
\item{data}{The database.}

\item{main.var}{string. The categorical variable against which each variable
will be compared.}

\item{var}{vector of strings. The categorical variables to compare with.}

\item{nb.dec}{digit. Number of decimals for the percentage.}
}
\value{
table
}
\description{
Performs the function \code{chisq.test} for the \code{main.var} with each
feature in \code{var}.
}
