% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CritCF_index_functions.R
\name{CritCF}
\alias{CritCF}
\title{\code{CritCF} index}
\usage{
CritCF(data, k, method, distance, Seed = 1)
}
\arguments{
\item{data}{dataframe for which the number of cluster should be estimated.}

\item{k}{integer, number of clusters.}

\item{method}{string, clustering algorithm to use. Available values are
"\code{kmeans}", "\code{hc}" (for \code{hclust()}) or "\code{mclust}".}

\item{distance}{string, distance between the observations (either "euclidean"
or "manhattan").}

\item{Seed}{If not \code{null}, passed to \code{set.seed()} before generating
the partition.}
}
\value{
a list, containing the criterion value and the partition
}
\description{
Computes \code{CritCF} index for the selected number of clusters and
clustering algorithm.
}
\keyword{internal}
