% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CritCF_index_functions.R
\name{CritCF.sel}
\alias{CritCF.sel}
\title{Number of cluster selection according to CritCF index}
\usage{
CritCF.sel(data, min.nc, max.nc, method, distance)
}
\arguments{
\item{data}{dataframe for which the number of cluster should be estimated.}

\item{min.nc}{interger strictly higher than 1 : minimum number of clusters.}

\item{max.nc}{interger (higher than min.nc) : maximum number of clusters.}

\item{method}{clustering algorithm to use.}

\item{distance}{distance between the observations (either euclidean or manhattan).}
}
\value{
A list containing the selected number of clusters, the CritCF values
and the best partition.
}
\description{
Wrapper to estimate the best number of clusters according to CritCF index
}
\examples{
doMIsaul:::CritCF.sel(iris[, 1:4], 2, 9, "kmeans", "euclidean")
}
