% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVE1_basic.R
\name{CVE1_basic}
\alias{CVE1_basic}
\title{Cross-valitation for cox regression using the basic approach}
\usage{
CVE1_basic(data, partition, nfolds = 10)
}
\arguments{
\item{data}{data contaning columns time and status}

\item{partition}{dataframe with 1 colunm}

\item{nfolds}{number of fold for CV}
}
\value{
numeric, cross-validation error
}
\description{
Cross-valitation for cox regression using the basic approach
}
\examples{
part <- data.frame(x= as.numeric(iris[, 5]))
CVE1_basic(iris[, 1:4],part, 10)
}
