% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVE2_VandVH.R
\name{CVE1_VandVH}
\alias{CVE1_VandVH}
\title{Cross-valitation for cox regression using Verweij and Van Houwelingen method}
\usage{
CVE1_VandVH(data, partition, nfolds = 10)
}
\arguments{
\item{data}{data contaning columns time and status}

\item{partition}{dataframe with 1 colunm}

\item{nfolds}{number of fold for CV}
}
\value{
numeric, cross-validation error
}
\description{
Cross-valitation for cox regression using Verweij and Van Houwelingen method
}
