% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVE1_basic.R
\name{cve_bas}
\alias{cve_bas}
\title{Cross-validation for cox regression using the basic approach}
\usage{
cve_bas(data, partition, nfolds = 10)
}
\arguments{
\item{data}{data containing columns time and status.}

\item{partition}{dataframe with 1 column.}

\item{nfolds}{number of fold for CV.}
}
\value{
numeric, cross-validation error
}
\description{
Cross-validation for cox regression using the basic approach
}
\examples{
data(cancer, package = "survival")
part <- data.frame(Cl= factor(cancer[, "sex"]), stringsAsFactors = T)
doMIsaul:::cve_bas(cancer,part, 10)
}
