% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVE3_LinearPred.R
\name{CVE_LinearPred}
\alias{CVE_LinearPred}
\title{Cross-valitation for cox regression using the linear predicor estimator}
\usage{
CVE_LinearPred(data, partition, nfolds = 10)
}
\arguments{
\item{data}{data contaning columns time and status}

\item{partition}{dataframe with 1 colunm}

\item{nfolds}{number of fold for CV}
}
\value{
numeric, cross-validation error
}
\description{
Cross-valitation for cox regression using the linear predicor estimator
}
