% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVE3_LinearPred.R
\name{CVE_LP}
\alias{CVE_LP}
\title{Cross-valitation for cox regression using the linear predicor estimator with
wrapper for warnings handling}
\usage{
CVE_LP(x)
}
\arguments{
\item{x}{list of 3 named elements : data (data contaning columns time and
status), partition (dataframe with 1 colunm), nfolds (number of fold for
CV)}
}
\value{
numeric, cross-validation error
}
\description{
Cross-valitation for cox regression using the linear predicor estimator with
wrapper for warnings handling
}
\examples{
data(cancer, package = "survival")
cancer$status <- cancer$status - 1
part <- data.frame(Cl= factor(cancer[, "sex"]), stringsAsFactors = T)
CVE_LP(list(data = cancer, partition = part, nfolds = 10))
}
