% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVE2_VandVH.R
\name{CVE_VandVH}
\alias{CVE_VandVH}
\title{Cross-valitation for cox regression using Verweij and Van Houwelingen
estimator with wrapper for warnings handling}
\usage{
CVE_VandVH(x)
}
\arguments{
\item{x}{list of 3 named elements : data (data contaning columns time and
status), partition (dataframe with 1 colunm), nfolds (number of fold for
CV)}
}
\value{
numeric, cross-validation error
}
\description{
Cross-valitation for cox regression using Verweij and Van Houwelingen
estimator with wrapper for warnings handling
}
\examples{
data(cancer, package = "survival")
part <- data.frame(Cl= factor(cancer[, "sex"]), stringsAsFactors = T)
CVE_VandVH(list(data = cancer, partition = part, nfolds = 10))
}
