% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVE2_VandVH.R
\name{cve_VVH}
\alias{cve_VVH}
\title{Cross-validation for cox regression using Verweij and Van Houwelingen method}
\usage{
cve_VVH(data, partition, nfolds = 10)
}
\arguments{
\item{data}{data containing columns time and status.}

\item{partition}{dataframe with 1 column.}

\item{nfolds}{number of fold for CV.}
}
\value{
numeric, cross-validation error
}
\description{
Cross-validation for cox regression using Verweij and Van Houwelingen method
}
\examples{
data(cancer, package = "survival")
part <- data.frame(Cl= factor(cancer[, "sex"]), stringsAsFactors = TRUE)
doMIsaul:::cve_VVH(cancer, part, 10)
}
