% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_imputers_censored.R
\name{.cens.draw3}
\alias{.cens.draw3}
\title{Base function for imputing left censored data with MICE}
\usage{
.cens.draw3(y, ry, x, lod.j = lod.j, ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the the subset
\code{y[ry]} of elements in y to which the imputation model is fitted.
The \code{ry} generally distinguishes the observed (\code{TRUE}) and
missing values (\code{FALSE}) in y.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors
for \code{y}. Matrix \code{x} may have no missing values.}

\item{lod.j}{censoring value.}

\item{...}{Other named arguments.}
}
\value{
parameters
}
\description{
from \code{Lapidus et al.}
}
\keyword{internal}
