% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_AUC.R
\name{Extract_AUC}
\alias{Extract_AUC}
\title{Wrapper to evaluate time dependent AUC}
\usage{
Extract_AUC(data, predictor, time.vect)
}
\arguments{
\item{data}{dataframe containing time and status information, as well as
marker}

\item{predictor}{\code{colname} of the marker to use to calculate AUC}

\item{time.vect}{vector of time points to consider}
}
\value{
AUC at \code{max(time.vect)} and median AUC along \code{time.vect}.
}
\description{
Wrapper to evaluate time dependent AUC
}
\keyword{internal}
