% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatpv.R
\name{formatpv}
\alias{formatpv}
\title{Apply format to pvalues}
\usage{
formatpv(p, text = FALSE)
}
\arguments{
\item{p}{digit. The pvalue to format}

\item{text}{boolean. Set to \code{TRUE}to display \code{"p="}, to
\code{FALSE} to display only the value.}
}
\value{
pvalue
}
\description{
Apply format to pvalues
}
