% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIclust_mpool.R
\name{MIclust_mpool}
\alias{MIclust_mpool}
\title{\code{MultiCons} wrapper for imputed datasets}
\usage{
MIclust_mpool(list.part, comb.cons, mcons.JAC.sel = 0, plot.MIclust = FALSE)
}
\arguments{
\item{list.part}{list of partitions with one element of the list corresponds
to the results of one imputed dataset. If more than one algorithm were
tester, each element if the list is a dataframe, as obtained by
\code{partition_generation()}.}

\item{comb.cons}{logical. Forced to \code{FALSE} if \code{length(algo)<2}.
Use \code{TRUE} to perform an additional consensus from all partitions.}

\item{mcons.JAC.sel}{passed to internal function \code{my_jack()}. Minimum
\code{Jaccard index} value between partitions to keep them for the
consensus.}

\item{plot.MIclust}{should \code{MultiCons} tree be plotted?}
}
\value{
a data frame with \code{ncol} = number of algorithms (+1 if
\code{comb.cons == T}), containing the consensus partitions.
}
\description{
Performs \code{MultiCons} from a list of partitions
}
\examples{
list.partitions <- list(
        partition_generation(iris[, 1:4], LOG = FALSE,
        clust.algo = c("km", "hc"),
        k.crit = "ch"),
        partition_generation(iris[, 1:2], LOG = FALSE,
        clust.algo = c("km", "hc"),
        k.crit = "ch"),
        partition_generation(iris[, 2:4], LOG = FALSE,
        clust.algo = c("km", "hc"),
        k.crit = "ch"))

doMIsaul:::MIclust_mpool(list.partitions, FALSE,0, FALSE)
}
