% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MImpute_surv.R
\name{MImpute_surv}
\alias{MImpute_surv}
\title{Wrapper function for multivariate imputation with survival data}
\usage{
MImpute_surv(
  data,
  mi.m,
  time.status.names = c("time", "status"),
  return.midsObject = FALSE
)
}
\arguments{
\item{data}{Dataframe with incomplete data.}

\item{mi.m}{Number of imputations to perform.}

\item{time.status.names}{Names of the variables for time ans staus (in that order)}

\item{return.midsObject}{Boolean}
}
\value{
If \code{return.midsObject == FALSE} a list of size mi.m, containing
the imputed datasets. If \code{return.midsObject == TRUE} a list of 2, the
first element (\code{imputed.data}) being the list of size mi.m as
described in the previous sentence, the 2nd element (\code{mids.obj})
containing the mids object as returmed by \code{mice()}
}
\description{
Performs imputation of the missing data using MICE for a dataset with
survival data. The Nelson Aalen estiamtor is calculated and used as predictor
in the impuation, Time is not used as predictor.
}
\examples{
DB <- XXX
MImpute_surv(DB, 10)
}
