% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCons.R
\name{MultiCons}
\alias{MultiCons}
\title{MultiCons Consensus Clustering Algorithm}
\usage{
MultiCons(
  DB,
  Clust_entry = FALSE,
  Clustering_selection = c("kmeans", "pam", "OPTICS", "agghc", "AGNES", "DIANA",
    "MCLUST", "CMeans", "FANNY", "BaggedClust"),
  num_algo = 10,
  maxClust = 10,
  sim.indice = "jaccard",
  returnAll = FALSE,
  Plot = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{DB}{Either data or dataframe of partitions.}

\item{Clust_entry}{Is DB partitions (T) or data (F).}

\item{Clustering_selection}{If DB is data, clustering algorithm to select
among. Must be included in default value.}

\item{num_algo}{Number of clustering algorithms to perform.}

\item{maxClust}{Maximum number of clusters.}

\item{sim.indice}{Index for defining best partition.}

\item{returnAll}{Should all partitions (T) or only the best (F) be returned.}

\item{Plot}{Should tree be plotted.}

\item{verbose}{Passed on to mclust and other fucntions.}
}
\value{
A list of 2 : performances and partitions. If returnAll is TRUE, both
elements of the list contain resutlts for all lelevs of the tree, else they
only ontain the results for  the bst level of the tree.
}
\description{
Performs Multicons clustering, from Al-Najdi et Al.
}
\examples{
# TO DO
}
