% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncvsurv_modified_functions.R
\name{my.cv.ncvsurv}
\alias{my.cv.ncvsurv}
\title{Modified \code{ncvreg:::cv.ncvsurv}.}
\usage{
my.cv.ncvsurv(
  X,
  y,
  ...,
  cluster,
  nfolds = 10,
  seed,
  fold,
  se = c("quick", "bootstrap"),
  returnY = FALSE,
  trace = FALSE
)
}
\arguments{
\item{X}{See \code{ncvreg:::cv.ncvsurv}.}

\item{y}{See \code{ncvreg:::cv.ncvsurv}.}

\item{...}{See \code{ncvreg:::cv.ncvsurv}.}

\item{cluster}{See \code{ncvreg:::cv.ncvsurv}.}

\item{nfolds}{See \code{ncvreg:::cv.ncvsurv}.}

\item{seed}{See \code{ncvreg:::cv.ncvsurv}.}

\item{fold}{See \code{ncvreg:::cv.ncvsurv}.}

\item{se}{See \code{ncvreg:::cv.ncvsurv}.}

\item{returnY}{See \code{ncvreg:::cv.ncvsurv}.}

\item{trace}{See \code{ncvreg:::cv.ncvsurv}.}
}
\description{
Modification of the \code{cv.ncvsurv()} function of \code{ncvreg} to use the
modified functions \code{ncvsurv()} : \code{my.ncvsurv()} and
\code{loss.ncvsurv()} : \code{mylossncvsurv()}.
}
