% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncvsurv_modified_functions.R
\name{my.ncvsurv}
\alias{my.ncvsurv}
\title{Modified \code{ncvreg:::ncvsurv}.}
\usage{
my.ncvsurv(
  X,
  y,
  penalty = c("MCP", "SCAD", "lasso"),
  gamma = switch(penalty, SCAD = 3.7, 3),
  alpha = 1,
  lambda.min = ifelse(n > p, 0.001, 0.05),
  nlambda = 100,
  lambda,
  eps = 1e-05,
  max.iter = 10000,
  convex = TRUE,
  dfmax = p,
  penalty.factor = rep(1, ncol(X)),
  warn = TRUE,
  returnX,
  ...
)
}
\arguments{
\item{X}{See \code{ncvreg:::ncvsurv}.}

\item{y}{See \code{ncvreg:::ncvsurv}.}

\item{penalty}{See \code{ncvreg:::ncvsurv}.}

\item{gamma}{See \code{ncvreg:::ncvsurv}.}

\item{alpha}{See \code{ncvreg:::ncvsurv}.}

\item{lambda.min}{See \code{ncvreg:::ncvsurv}.}

\item{nlambda}{See \code{ncvreg:::ncvsurv}.}

\item{lambda}{See \code{ncvreg:::ncvsurv}.}

\item{eps}{See \code{ncvreg:::ncvsurv}.}

\item{max.iter}{See \code{ncvreg:::ncvsurv}.}

\item{convex}{See \code{ncvreg:::ncvsurv}.}

\item{dfmax}{See \code{ncvreg:::ncvsurv}.}

\item{penalty.factor}{See \code{ncvreg:::ncvsurv}.}

\item{warn}{See \code{ncvreg:::ncvsurv}.}

\item{returnX}{See \code{ncvreg:::ncvsurv}.}

\item{...}{See \code{ncvreg:::ncvsurv}.}
}
\description{
modification of the \code{ncvsurv()} function (\code{ncvreg} internal)
with removal of the condition \code{nlambda > 2}.
}
