% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_jack.R
\name{my_jack}
\alias{my_jack}
\title{Partition sorting based on Jaccard index}
\usage{
my_jack(data, mcons.JAC.sel = 0)
}
\arguments{
\item{data}{dataframe containing the partitions (factors)}

\item{mcons.JAC.sel}{minimum required Jaccard index between partitions (0 =
all partitions kept)}
}
\value{
dataframe of partitions with only the partitions satisfying the
conditions kept.
}
\description{
Evaluates if all partitions are the same (and thus no consensus algorithm
needed) and may pre-select partitions depending on a threshold on the Jaccard
index
}
\examples{
partitions <- data.frame(factor(rep(c(1, 2), each = 100)),
                         factor(rep(c(1, 2), times = c(120, 80))),
                         factor(rep(c(2, 1), each = 100)),
                         factor(rep(c(1, 2, 3, 4), each = 50)))
dim(doMIsaul:::my_jack(partitions, 0))
dim(doMIsaul:::my_jack(partitions, 0.5))
}
\keyword{internal}
