% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncvsurv_modified_functions.R
\name{mylossncvsurv}
\alias{mylossncvsurv}
\title{Modified \code{ncvreg:::loss.ncvsurv}.}
\usage{
mylossncvsurv(y, eta, total = TRUE)
}
\arguments{
\item{y}{See \code{ncvreg:::loss.ncvsurv}.}

\item{eta}{See \code{ncvreg:::loss.ncvsurv}.}

\item{total}{See \code{ncvreg:::loss.ncvsurv}.}
}
\value{
See \code{ncvreg:::loss.ncvsurv}.
}
\description{
modification of the \code{loss.ncvsurv()} function (\code{ncvreg} internal)
to handle a single value of lambda
}
\keyword{internal}
