% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto}
\alias{pareto}
\title{Pareto optimisation}
\usage{
pareto(bi.objective, obj.names = c("Cluster", "Regression"))
}
\arguments{
\item{bi.objective}{dataframe or matrix containing the objective values for
each case}

\item{obj.names}{name of the twoobjectives - should be in
colname(bi.objetive)}
}
\value{
the entry bi.objective with an additional column "Domimanted" with 1
for non optimal solutions and 0 for optimal.
}
\description{
Find pareto optimal and pareto non optimal solutions
}
\examples{
tab <- data.frame(A = rnorm(100, 10, 3), B = runif(100, 3,4))
doMIsaul:::pareto(tab, c("A", "B"))
}
