% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_generation.R
\name{partition_generation}
\alias{partition_generation}
\title{Unsupervised partition with K selection}
\usage{
partition_generation(data, LOG, clust.algo, k.crit)
}
\arguments{
\item{data}{dataframe to cluster}

\item{LOG}{logical. Should all columns of the dataset be logged before
applying clustering algorithms?}

\item{clust.algo}{vector of strings: name of clsutering algorithms to use
(use "km" for k-means, "kmed" for K-medians, "hc" for hclust and/or "mclust" for mclust).}

\item{k.crit}{string. Criterion to select the optimal number of clusters (for
each imputed dataset). Use "ch" for Calinski and Harabasz criterion (not
available for mclust), "CritCF" for CriCF or bic for BIC (mclust only).}
}
\value{
a dataframe with one column for each algortihm in clust.algo ,
containing the cluster IDs.
}
\description{
Generates a partition using clust.algo algorithm, with k.crit for selecting
the number of clusters
}
\examples{
partition_generation(iris[, 1:4], LOG = F, clust.algo = c("km", "hc"), k.crit = "ch")
}
