% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MIpca.R
\name{plot_MIpca}
\alias{plot_MIpca}
\alias{plot_MIpca_all}
\title{plot a PCA from a multiply imputed dataset.}
\usage{
plot_MIpca(data.list, obs.sel = NULL, pca.varsel = NULL, pc.sel = c(1, 2))

plot_MIpca_all(data.list, obs.sel = NULL, pca.varsel = NULL, pc.sel = c(1, 2))
}
\arguments{
\item{data.list}{list of the imputed datasets.}

\item{obs.sel}{selection of observations to highlight}

\item{pca.varsel}{optional. A variables selection for the PCA}

\item{pc.sel}{which PC to draw. default is 1 and 2}
}
\value{
\code{ggplot} object
}
\description{
\code{plot_MIpca} draws only mean value while \code{plot_MIpca_all} draws all
values for the selected observations.
}
\examples{
data(cancer, package = "survival")
cancer.imp <- MImpute(cancer[, -c(1:3)], 6)
plot_MIpca(cancer.imp, 1:10,
    pca.varsel = c("age", "sex", "ph.ecog", "meal.cal",  "wt.loss"))
data(cancer, package = "survival")
cancer.imp <- MImpute(cancer[, -c(1:3)], 6)
plot_MIpca_all(cancer.imp, 1:10,
    pca.varsel = c("age", "sex", "ph.ecog", "meal.cal",  "wt.loss"))
}
