% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupMI.R
\name{unsupMI}
\alias{unsupMI}
\title{Unsupervised learning for incomplete dataset}
\usage{
unsupMI(
  Impute = FALSE,
  Impute.m = 5,
  cens.data.lod = NULL,
  cens.standards = NULL,
  cens.mice.log = 10,
  data,
  log.data = FALSE,
  algo = "km",
  k.crit = "ch",
  comb.cons = FALSE,
  plot.cons = FALSE,
  return.detail = FALSE,
  not.to.use = c("time", "status")
)
}
\arguments{
\item{Impute}{Default is FALSE to indicate that the user performed
the imputation and provides the imputed data. Otherwise string
("\code{MImpute}", "\code{MImpute_surv}" or "\code{MImpute_lcens}") to
perform the imputation within the call using the \code{MImpute()},
\code{MImpute_surv()} or \code{MImpute_lcens()} function.}

\item{Impute.m}{Used only if Impute is not \code{FALSE} ; number of
imputations to perform}

\item{cens.data.lod}{passed to \code{MImpute_lcens()} if
\code{Impute == MImpute_lcens}}

\item{cens.standards}{passed to \code{MImpute_lcens()} if
\code{Impute == MImpute_lcens}}

\item{cens.mice.log}{passed to \code{MImpute_lcens()} if
\code{Impute == MImpute_lcens}}

\item{data}{Data, in the form of a list of data.frame(s). The list should be
one length 1 if data are complete or if Impute is not \code{FALSE}, it
should be a list of imputed dataframes if data are incomplete and imputed.
If columns named "\code{time}" and "\code{status}" are present they will be
discarded for the clustering.}

\item{log.data}{logical. Should all columns of the dataset be logged before
applying clustering algorithms?}

\item{algo}{vector of strings: name of clustering algorithms to use (use
"\code{km}" for \code{k-means}, "\code{kmed}" for \code{K-medians},
"\code{hc}" for \code{hclust} and/or "\code{mclust}" for \code{mclust}).}

\item{k.crit}{string. Criterion to select the optimal number of clusters (for
each imputed dataset). Use "\code{ch}" for Calinski and Harabasz criterion
(not available for \code{mclust}), "\code{CritCF}" for \code{CriCF} or
"\code{bic}" for \code{BIC} (\code{mclust} only).}

\item{comb.cons}{logical. Forced to \code{FALSE} if \code{length(algo)<2}.
Use \code{TRUE} to perform an additional consensus from all partitions
generates, whatever the algorithm.}

\item{plot.cons}{logical. Use \code{TRUE} to print the MultiCons tree.}

\item{return.detail}{logical. Should the detail of imputation specific
partition be returned, in the supplement to the final consensus partition?}

\item{not.to.use}{vector of strings : names of the columns that should be
discarded for the learning step.}
}
\value{
if \code{length(algo)>1} a vector of final cluster ID ; if
\code{length(algo)>1} a data.frame with each column being the final cluster
ID for the corresponding algorithm.
}
\description{
Unsupervised clustering for multiply imputed datasets using \code{MultiCons}
consensus (Faucheux et al. 2021 procedure)
}
\examples{
## With imputation included TO DO
#
### With imputation not included
## 1 imputation
data(cancer, package = "survival")
cancer$status <- cancer$status - 1
cancer.imp <- MImpute_surv(cancer, 3)
## 2 learning
res <- unsupMI(data = cancer.imp)
## 2.bis learning with several algorithms
res.2 <- unsupMI(data = cancer.imp, algo = c("km", "hc"), comb.cons = TRUE,
                 plot.cons = TRUE)
## Alternative: perform imputation within
res <- unsupMI(Impute = "MImpute_surv", data = list(cancer))
}
