% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allocation_Distance.R
\name{Allocation_Distance}
\alias{Allocation_Distance}
\title{Allocation Distance}
\usage{
Allocation_Distance(Z, Xr)
}
\arguments{
\item{Z}{dataset of dimension \code{n * p}.}

\item{Xr}{Center coordinates of dimension \code{k * p}.}
}
\value{
List with distances to center and allocation matrix
}
\description{
\code{Khorshidi et al., 2019}.
}
\keyword{internal}
