% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH_index_functions.R
\name{CH}
\alias{CH}
\title{\code{CH} index}
\usage{
CH(data, k, method, Seed = 1)
}
\arguments{
\item{data}{dataframe for which the number of cluster should be estimated.}

\item{k}{integer, number of clusters.}

\item{method}{string, clustering algorithm to use. The only available value
for now is "\code{kmed}" for \code{K-medians} clustering.}

\item{Seed}{If not \code{null}, passed to \code{set.seed()} before generating
the partition.}
}
\value{
a list, containing the criterion value and the partition.
}
\description{
Computes CH index for the selected number of clusters and clustering
algorithm.
}
\keyword{internal}
