% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH_index_functions.R
\name{CH.sel}
\alias{CH.sel}
\title{Number of cluster selection according to \code{CH} index}
\usage{
CH.sel(data, min.nc, max.nc, method)
}
\arguments{
\item{data}{dataframe for which the number of cluster should be estimated.}

\item{min.nc}{integer strictly higher than 1: minimum number of clusters.}

\item{max.nc}{integer (\code{>min.nc}): maximum number of  clusters.}

\item{method}{string, clustering algorithm to use. The only available values
for now are "\code{kmed}" for \code{K-medians} clustering and
"\code{kproto}" for \code{clustMixType::kproto()} clustering.}
}
\value{
A list containing the selected number of clusters, the CH values and
the best partition.
}
\description{
Wrapper to estimate the best number of clusters according to \code{CH} index.
}
\keyword{internal}
