% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2tab.R
\name{chi2tab}
\alias{chi2tab}
\title{Table of \code{chisq.test()} test for several explanatory variables}
\usage{
chi2tab(data, main.var, var, nb.dec = 1)
}
\arguments{
\item{data}{The database.}

\item{main.var}{string. The categorical variable against the variable
will be compared.}

\item{var}{string. The categorical variable to compare with.}

\item{nb.dec}{digit. Number of decimals for the percentage.}
}
\value{
table with n (percentage) per level and p-value of \code{chi2} test.
}
\description{
Performs the function \code{chisq.test()} for the \code{main.var} and
\code{var}.
}
\keyword{internal}
