% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVE1_basic.R
\name{CVE_basic}
\alias{CVE_basic}
\title{Cross-validation for cox regression using the basic approach
estimator with wrapper for warnings handling}
\usage{
CVE_basic(x)
}
\arguments{
\item{x}{list of 3 named elements : data (data containing columns time and
status), partition (dataframe with 1 column), nfolds (number of fold for
CV)?}
}
\value{
numeric, cross-validation error
}
\description{
Cross-validation for cox regression using the basic approach
estimator with wrapper for warnings handling
}
\examples{
data(cancer, package = "survival")
part <- data.frame(Cl= factor(cancer[, "sex"]), stringsAsFactors = T)
CVE_basic(list(data = cancer, partition = part, nfolds = 10))
}
