% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVE1_basic.R
\name{CVE_basic}
\alias{CVE_basic}
\title{Cross-valitation for cox regression using the basic approach
estimator with wrapper for warnings handling}
\usage{
CVE_basic(x)
}
\arguments{
\item{x}{list of 3 named elements : data (data contaning columns time and
status), partition (dataframe with 1 colunm), nfolds (number of fold for
CV)}
}
\value{
numeric, cross-validation error
}
\description{
Cross-valitation for cox regression using the basic approach
estimator with wrapper for warnings handling
}
\examples{
part <- data.frame(x= as.numeric(iris[, 5]))
CVE_basic(list(data = iris[, 1:4], partition = part, nfolds = 10))
}
