% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVE3_LinearPred.R
\name{cve_LinearPred}
\alias{cve_LinearPred}
\title{Cross-valitation for cox regression using the linear predicor estimator}
\usage{
cve_LinearPred(data, partition, nfolds = 10)
}
\arguments{
\item{data}{data contaning columns time and status}

\item{partition}{dataframe with 1 colunm}

\item{nfolds}{number of fold for CV}
}
\value{
numeric, cross-validation error
}
\description{
Cross-valitation for cox regression using the linear predicor estimator
}
\examples{
data(cancer, package = "survival")
cancer$status <- cancer$status - 1
part <- data.frame(Cl= factor(cancer[, "sex"]), stringsAsFactors = T)
doMIsaul:::cve_LinearPred(cancer,part, 10)
}
