% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVE3_LinearPred.R
\name{cve_LinearPred}
\alias{cve_LinearPred}
\title{Cross-validation for cox regression using the linear predictor estimator}
\usage{
cve_LinearPred(data, partition, nfolds = 10)
}
\arguments{
\item{data}{data containing columns time and status.}

\item{partition}{dataframe with 1 column containing the partition labels.}

\item{nfolds}{number of fold for CV.}
}
\value{
numeric, cross-validation error
}
\description{
Cross-validation for cox regression using the linear predictor estimator
}
\keyword{internal}
