% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exctract_center_position.R
\name{exctract_center_position}
\alias{exctract_center_position}
\title{Extract the cluster centers coordinates}
\usage{
exctract_center_position(data, Cluster, centPos)
}
\arguments{
\item{data}{dataframe (of dimension \code{n * p}).}

\item{Cluster}{vector (of length \code{n}) containing the cluster
allocations.}

\item{centPos}{String, indicated whether the center position should be
defined by the mean (use "\code{colMeans}") or the median of the cluster
(use any other string).}
}
\value{
matrix (\code{k * p}) with center coordinates
}
\description{
Extract the cluster centers coordinates
}
\keyword{internal}
