% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIclust_mpool.R
\name{MIclust_mpool}
\alias{MIclust_mpool}
\title{Multicons wrapper for imputed datasets}
\usage{
MIclust_mpool(list.part, comb.cons, plot.MIclust = FALSE)
}
\arguments{
\item{list.part}{list of partitions with one element of the list corresponds
to the results of one imputed dataset. If more than one algorithm were
tester, each element if the list is a dataframe, as obtained by
partition_generation()}

\item{comb.cons}{logical. Forced to FALSE if \code{length(algo)<2}. Use
\code{TRUE} to perform an additional consensus from all partitions}

\item{plot.MIclust}{should Multicons tree be plotted?}
}
\value{
a data frame with ncol = number of algorithms (+1 if comb.cons ==
T), containing the consensus partitions
}
\description{
Performs Multicons from a list of partitions
}
\examples{
list.partitions <- partition_generation(iris[, 1:4], LOG = FALSE,
                                        clust.algo = c("km", "hc"),
                                        k.crit = "ch")
# MIclust_mpool(list.partitions, F, F)
}
