% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_modified_function.R
\name{my.cv.glmnet}
\alias{my.cv.glmnet}
\title{Modified \code{glmnet:::cv.glmnet}.}
\usage{
my.cv.glmnet(
  x,
  y,
  weights = NULL,
  offset = NULL,
  lambda = NULL,
  type.measure = c("default", "mse", "deviance", "class", "auc", "mae", "C"),
  nfolds = 10,
  foldid = NULL,
  alignment = c("lambda", "fraction"),
  grouped = TRUE,
  keep = FALSE,
  parallel = FALSE,
  gamma = c(0, 0.25, 0.5, 0.75, 1),
  relax = FALSE,
  trace.it = 0,
  ...
)
}
\arguments{
\item{x}{See \code{glmnet:::cv.glmnet}.}

\item{y}{See \code{glmnet:::cv.glmnet}.}

\item{weights}{See \code{glmnet:::cv.glmnet}.}

\item{offset}{See \code{glmnet:::cv.glmnet}.}

\item{lambda}{See \code{glmnet:::cv.glmnet}.}

\item{type.measure}{See \code{glmnet:::cv.glmnet}.}

\item{nfolds}{See \code{glmnet:::cv.glmnet}.}

\item{foldid}{See \code{glmnet:::cv.glmnet}.}

\item{alignment}{See \code{glmnet:::cv.glmnet}.}

\item{grouped}{See \code{glmnet:::cv.glmnet}.}

\item{keep}{See \code{glmnet:::cv.glmnet}.}

\item{parallel}{See \code{glmnet:::cv.glmnet}.}

\item{gamma}{See \code{glmnet:::cv.glmnet}.}

\item{relax}{See \code{glmnet:::cv.glmnet}.}

\item{trace.it}{See \code{glmnet:::cv.glmnet}.}

\item{...}{See \code{glmnet:::cv.glmnet}.}
}
\description{
Modification of the \code{cv.glmnet} function (\code{glmnet}
package version 2.0-18)
with removal of the condition \code{lenght(lamba) > 2}
}
