% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_jack.R
\name{my_jack}
\alias{my_jack}
\title{Partition sorting based on Jaccard index}
\usage{
my_jack(data, mcons.JAC.sel = 0)
}
\arguments{
\item{data}{dataframe containing the partitions (factors). Specifically, each
column must be a partition and must be a factor.}

\item{mcons.JAC.sel}{Numeric (in (0, 1)), minimum \code{Jaccard index} value
between partitions to keep them (0 = all partitions kept).}
}
\value{
dataframe of partitions with only the partitions satisfying the
conditions kept.
}
\description{
Evaluates if all partitions are the same (and thus no consensus algorithm
needed) and may pre-select partitions depending on a threshold on the Jaccard
index
}
\keyword{internal}
