% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_clustering.R
\name{objective_clustering}
\alias{objective_clustering}
\title{objective clustering cost}
\usage{
objective_clustering(Alloc_Dist)
}
\arguments{
\item{Alloc_Dist}{list with Distance to cluster centers and allocation
matrix, as obtained with \code{Allocation_Distance()}.}
}
\value{
matrix
}
\description{
Auxiliary function to calculate a cost function for clustering. Modified
version of \code{Cost_func()} from \code{Khorshidi et al., 2019}.
}
\keyword{internal}
