% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupMI.R
\name{unsupMI}
\alias{unsupMI}
\title{Unsupervised learning for incomplete dataset}
\usage{
unsupMI(data, log.data = FALSE, algo = "km", k.crit = "ch",
   comb.cons = FALSE, plot.cons = FALSE, return.detail = FALSE,
   not.to.use = c("time", "status"))
}
\arguments{
\item{data}{list of the imupted datasets (if present, columns names "time"
and "status" will be discared)}

\item{log.data}{logical. Should all columns of the dataset be logged before
applying clustering algorithms?}

\item{algo}{vector of strings: name of clsutering algorithms to use (use "km"
for k-means, "kmed" for K-medians, "hc" for hclust and/or "mclust" for
mclust).}

\item{k.crit}{string. Criterion to select the optimal number of clusters (for
each imputed dataset). Use "ch" for Calinski and Harabasz criterion (not
available for mclust), "CritCF" for CriCF or bic for BIC (mclust only).}

\item{comb.cons}{logical. Forced to FALSE if \code{length(algo)<2}. Use
\code{TRUE} to perform an additional consensus from all partitions
generates, whaterver the algorithm.}

\item{plot.cons}{logical. Use \code{TRUE} to print the MultiCons tree.}

\item{return.detail}{logical. Should the detail of imputation specific
partition be returned, in the supplument to the final consensus partition?}

\item{not.to.use}{vector of strings : names of the columns that should be
discraded for the learning step.}
}
\value{
if \code{length(algo)>1} a vector of final cluster ID ; if
\code{length(algo)>1} a data.frame with each column being the final cluster
ID for the corresponding algorithm.
}
\description{
Unsupervised clustering for multiply imputed datasets unsing MultiCons
consensus (Faucheux et al. 2020 procedure)
}
\examples{
# unsupMI()
}
